compose::implementation::library::diag::compose_codespan_reporting

Module term

Terminal back-end for emitting diagnostics.

Modules

termcolor

This crate provides a cross platform abstraction for writing colored text to a terminal. Colors are written using either ANSI escape sequences or by communicating with a Windows console. Much of this API was motivated by use inside command line applications, where colors or styles can be configured by the end user and/or the environment.

Structs

Chars

Characters to use when rendering the diagnostic.

ColorArg

A command line argument that configures the coloring of the output.

Config

Configures how a diagnostic is rendered.

Styles

Styles to use when rendering the diagnostic.

Enums

DisplayStyle

The display style to use when rendering diagnostics.

Functions

emit

Emit a diagnostic using the given writer, context, config, and files.