compose::implementation::syntax

Struct SyntaxNode

pub struct SyntaxNode(/* private fields */);

A node in the untyped syntax tree.

A node can either be

a leaf node representing a single token,

an inner node containing one or more children

or, an error node containing a syntax error.

Implementations

impl SyntaxNode

pub fn is<'a, T>(&'a self) -> bool
where T: AstNode<'a>,

pub fn cast<'a, T>(&'a self) -> Option<T>
where T: AstNode<'a>,

Tries to cast the node to type T. Returns None if it cannot.

pub fn try_cast_first<'a, T>(&'a self) -> Option<T>
where T: AstNode<'a>,

pub fn try_cast_last<'a, T>(&'a self) -> Option<T>
where T: AstNode<'a>,

impl SyntaxNode

pub fn kind(&self) -> SyntaxKind

pub fn text(&self) -> &EcoString

pub fn to_text(&self) -> EcoString

Like text, but also returns the text of inner nodes.

Builds a new string, so is more computationally expensive.

pub fn children(&self) -> Iter<'_, SyntaxNode>

pub fn span(&self) -> Span

impl SyntaxNode

pub fn errors(&self) -> Vec<SyntaxError>

pub fn warnings(&self) -> Vec<SyntaxError>

pub fn error_mut(&mut self) -> Option<&mut SyntaxError>