compose::language::Error_Codes

Module DOC_E0004_MUTATE_IMMUTABLE_VARIABLE

E0004: Cannot reassign to a variable declared as immutable

Variables declared with let are immutable by default.

✅ Write let mut if you want to change their value after initialization.

Example of erroneous code

let b = 4;
b = 5;

error[E0004]: cannot reassign to a variable declared as immutable
  ┌─ main.comp:2:1
  │
1 │ let b = 4;
  │     - was defined as immutable here
2 │ b = 5;
  │ ^ is immutable
  │
  = note: variables are immutable by default
  = help: make the variable mutable by writing `let mut`
  = help: for more information about this error, try `compose explain E0004`

Or

let a;
a = 2;
a = 3;

error[E0004]: cannot reassign to a variable declared as immutable
  ┌─ main.comp:3:1
  │
1 │ let a;
  │     - was defined as immutable here
2 │ a = 2;
  │ - first assignment occurred here
3 │ a = 3;
  │ ^ cannot reassign an immutable variable
  │
  = note: variables are immutable by default
  = help: make the variable mutable by writing `let mut`
  = help: for more information about this error, try `compose explain E0004`

Explanation

In Compose, variables are immutable by default. This means you can assign to them once, but attempting to reassign will produce an error.

To allow a variable to change over time, declare it as mutable using mut:

let mut counter = 0;
counter = counter + 1;

This design prevents accidental changes and helps promote safer and more predictable code.

Fixes

✅ Make the variable mutable:

let mut a = 2;
a = 3;

✅ Or, use shadowing (create a new variable):

let a = 2;
// ...
let a = 3; // this creates a new `a`

Shadowing can be a good alternative to mutation, especially in functional-style code.