compose::language::Error_Codes

Module DOC_E0301_ARRAY_DESTRUCTURING_WRONG_NUMBER_OF_ELEMENTS

E0301: Array destructuring pattern did not fit array length

This error occurs when an array destructuring pattern does not match the number of elements in the array.

Array destructuring in Compose requires that:

If the pattern does not use .., the number of bindings must exactly match the array’s length.

If the pattern uses a .. (rest) binding, the array must have at least as many elements as the non-rest bindings.

Violating these rules results in a length mismatch error.

Too many elements in the array, pattern without ..

let [a, b] = [1, 2, 3];

error[E0301]: too many elements to destructure
  ┌─ main.comp:1:5
  │
1 │ let [a, b] = [1, 2, 3];
  │     ^^^^^^
  │
  = help: the provided array has a length of 3, but the pattern expects 2 elements
  = help: use `..` to ignore the remaining elements, or `..name` to bind them
  = help: for more information about this error, try `compose explain E0301`

Pattern binds 2 elements.

Array has 3 elements.

Error: the pattern does not account for all elements.

Fix: Use .. to ignore or capture the remaining elements:

let [a, b, ..] = [1, 2, 3];       // ignore remaining
let [a, b, ..rest] = [1, 2, 3];   // capture remaining

Case 2: Too few elements in the array, pattern without ..

let [a, b, c] = [1, 2];

error[E0301]: not enough elements to destructure
  ┌─ main.comp:1:5
  │
1 │ let [a, b, c] = [1, 2];
  │     ^^^^^^^^^
  │
  = help: the provided array has a length of 2, but the pattern expects 3 elements
  = help: for more information about this error, try `compose explain E0301`

Pattern binds 3 elements.

Array has only 2 elements.

Error: not enough elements to satisfy the pattern.

Fix: Provide an array with enough elements:

let [a, b, c] = [1, 2, 3];  

Case 4: Too few elements when using a ..name binding

let [a, b, ..c] = [1];

error[E0301]: not enough elements to destructure
  ┌─ main.comp:1:5
  │
1 │ let [a, b, ..c] = [1];
  │     ^^^^^^^^^^^
  │
  = help: the provided array has a length of 1, but the pattern expects at least 2 elements
  = help: for more information about this error, try `compose explain E0301`

Pattern requires 2 non-rest elements (a and b).

Array has only 1 element.

Error: not enough elements to bind all required non-rest values.

Fix: Ensure the array has at least as many elements as the non-rest bindings:

let [a, b, ..c] = [1, 2];  

or reduce the number of non-rest bindings:

let [a, ..c] = [1];